use a txRef type with a isCancel indicator instead of a magic channel id indicator
