+++ op-e2e/system_test.go
@@ -1456,8 +1456,9 @@ func StopStartBatcher(t *testing.T, deltaTimeOffset *hexutil.Uint64) {
 	require.NoError(t, err)
 	require.Greater(t, newSeqStatus.SafeL2.Number, seqStatus.SafeL2.Number, "Safe chain did not advance")
 
+	driver := sys.BatchSubmitter.TestDriver()
 	// stop the batch submission
-	err = sys.BatchSubmitter.Driver().StopBatchSubmitting(context.Background())
+	err = driver.StopBatchSubmitting(context.Background())
 	require.NoError(t, err)
 
 	// wait for any old safe blocks being submitted / derived
@@ -1477,7 +1478,7 @@ func StopStartBatcher(t *testing.T, deltaTimeOffset *hexutil.Uint64) {
 	require.Equal(t, newSeqStatus.SafeL2.Number, seqStatus.SafeL2.Number, "Safe chain advanced while batcher was stopped")
 
 	// start the batch submission
-	err = sys.BatchSubmitter.Driver().StartBatchSubmitting()
+	err = driver.StartBatchSubmitting()
 	require.NoError(t, err)
 	time.Sleep(safeBlockInclusionDuration)
 
@@ -1519,7 +1520,8 @@ func TestBatcherMultiTx(t *testing.T) {
 	require.NoError(t, err)
 
 	// start batch submission
-	err = sys.BatchSubmitter.Driver().StartBatchSubmitting()
+	driver := sys.BatchSubmitter.TestDriver()
+	err = driver.StartBatchSubmitting()
 	require.NoError(t, err)
 
 	totalTxCount := 0
